Index des fonctions

Documentation de toute les fonctions du package CSAD

AnalyseDeDonnees.tp1_Composantes_principales

AnalyseDeDonnees.tp1_Donnees_centrees

AnalyseDeDonnees.tp1_Individu_moyen

AnalyseDeDonnees.tp1_Matrice_de_donnees

AnalyseDeDonnees.tp1_Matrice_var_cov

AnalyseDeDonnees.tp1_Sigma2

AnalyseDeDonnees.tp2_MCO

AnalyseDeDonnees.tp3_couleur_moyenne

AnalyseDeDonnees.tp3_estimation_mu_sigma

AnalyseDeDonnees.tp3_exercice1

AnalyseDeDonnees.tp3_exercice2

AnalyseDeDonnees.tp3_exercice3

AnalyseDeDonnees.tp3_exercice3bis

AnalyseDeDonnees.tp3_exercice4

AnalyseDeDonnees.tp3_moyenne_amelioree

AnalyseDeDonnees.tp3_vraisemblance

AnalyseDeDonnees.tp4_kppv

AnalyseDeDonnees.tp1_Composantes_principales — Method

Syntaxe

C = Composantes_principales(Im)

Entrée :

Im : (Array{RGBA{Normed{UInt8,8}},2}) matrice representatrice l'image

Sortie:

C : (Array{Float64,2}) la matrice qui contient les composantes principales en colonnes

AnalyseDeDonnees.tp1_Donnees_centrees — Method

Syntaxe

X_c = Donnees_centrees(Im)

Entrée :

Im : (Array{RGBA{Normed{UInt8,8}},2}) matrice representatrice l'image

Sortie:

X_c : (Array{Float64,2}) matrice des données centrés

AnalyseDeDonnees.tp1_Individu_moyen — Method

Syntaxe

x_barre = Individu_moyen(Im)

Entrée :

Im : (Array{RGBA{Normed{UInt8,8}},2}) matrice representatrice l'image

Sortie:

x_barre : (Array{Float64,2}) l'individu moyen

AnalyseDeDonnees.tp1_Matrice_de_donnees — Method

Syntaxe

X = Matrice_de_donnees(Im)

Entrée :

Im : (Array{RGBA{Normed{UInt8,8}},2}) matrice representatrice l'image

Sortie:

X : (Array{Float32,2}) la matrice des données

AnalyseDeDonnees.tp1_Matrice_var_cov — Method

Syntaxe

Sigma = Matrice_var_cov(Im)

Entrée :

Im : (Array{RGBA{Normed{UInt8,8}},2}) matrice representatrice l'image

Sortie:

Sigma : (Array{Float64,2}) la matrice de variance/covariance des variables aléatoires

AnalyseDeDonnees.tp1_Sigma2 — Method

Syntaxe

Sigma_2 = Sigma2(Im)

Entrée :

Im : (Array{RGBA{Normed{UInt8,8}},2}) matrice representatrice l'image

Sortie:

Sigma_2 : (Array{Float64,2}) Matrice de variance/covariance dans le nouveau repere

AnalyseDeDonnees.tp2_MCO — Method

Résolution approchée du problème au sens du moindre carrée

Syntaxe

X_sol = MCO(x,y)

Entrées :

x : (Array{Float,1}) le vecteur x definissant les paramètres du prb

y : (Array{Float,1}) le vecteur y definissant les paramètres du prb

Sorties:

X_sol : (Array{Float,1}) la solution du problème : AX=B

AnalyseDeDonnees.tp3_couleur_moyenne — Method

TP3 - Classification bayesienne

Calcule la couleur moyenne d'une image

Entrée

image : l'image

Sortie

[r_bar, v_bar] : les deux composantes moyenne (rouge & vert) de couleur normalisées $(\bar{r},\bar{v})$

AnalyseDeDonnees.tp3_estimation_mu_sigma — Method

TP3 - Classification bayesienne

L’estimation empirique des paramètres d’une loi normale bidimensionnelle à partir d’une matrice de données X.

Entrée

X : la matrice des donnees

Sorties

mu : l'espérance estimée $= E[X]$

sigma : la matrice de variance / covariance estimée $= E[(X - \mu)(X - \mu)^{T}]$

AnalyseDeDonnees.tp3_exercice1 — Method

TP3 - Classification bayesienne

Calcul de la couleur moyenne d’une image

Entrées

afficher : (Bool) afficher les figures (false dans le runtests.jl)

chemin : (String) le chemin vers ce fichier

AnalyseDeDonnees.tp3_exercice2 — Method

TP3 - Classification bayesienne

Estimation de la vraisemblance de chaque espèce de fleurs

Entrées

afficher : (Bool) afficher les figures (false dans le runtests.jl)

chemin : (String) le chemin vers ce fichier

AnalyseDeDonnees.tp3_exercice3 — Method

TP3 - Classification bayesienne

Classification d’images de fleurs sous l’hypothèse que les classes sont équiprobables

Entrées

afficher : (Bool) afficher les figures (false dans le runtests.jl)

chemin : (String) le chemin vers ce fichier

Sorties

accuracy : (Float64) l'accuracy du classifieur

AnalyseDeDonnees.tp3_exercice3bis — Method

TP3 - Classification bayesienne

Classification d’images de fleurs en ajustant les propbabilités a priori des classes

Entrées

afficher : (Bool) afficher les figures (false dans le runtests.jl)

chemin : (String) le chemin vers ce fichier

Sorties

accuracy : (Float64) l'accuracy du classifieur

AnalyseDeDonnees.tp3_exercice4 — Method

TP3 - Classification bayesienne

Amélioration du classifieur

Entrées

afficher : (Bool) afficher les figures (false dans le runtests.jl)

chemin : (String) le chemin vers ce fichier

Sorties

accuracy : (Float64) l'accuracy du classifieur

AnalyseDeDonnees.tp3_moyenne_amelioree — Method

TP3 - Classification bayesienne

Calcule les trois valeurs décrivant l'image $(\bar{r},\bar{v},\bar{r}_{C})$

Entrées

I : l'image

Sorties

[r_bar_P, v_bar_P, r_bar_C] : $(\bar{r},\bar{v})$ calculées sur le pourtour P et la valeur $\bar{r}_{C}$ calculée sur le centre C

AnalyseDeDonnees.tp3_vraisemblance — Method

TP3 - Classification bayesienne

calcule la vraisemblance $p(x|\omega_i)$ et le dénominateur de classe x

Entrées

x : les coordonnées sur la grille

mu_i : (Array{Float64,1}) l'espérance $\mu_{i}$

Sigma_i : (Array{Float64,2}) la matrice variance/covariance $\Sigma_{i}$

denominateur_classe_i : (Float64) le dénominateur de la i-ème classe

-1

ou fourni $= (2 \pi)^{d / 2}\left(\operatorname{det} \Sigma_{i}\right)^{1 / 2}$

Sorties

P : La vraisemblance $p(x|\omega_i)$

denominateur : (Float64) le denominateur de la i-ème classe $= (2 \pi)^{d / 2}\left(\operatorname{det} \Sigma_{i}\right)^{1 / 2}$

AnalyseDeDonnees.tp4_kppv — Method

TP4 - Reconnaissance de chiffres manuscrits par k plus proches voisins

Syntaxe

[Partition,confusion,nb_erreurs] = tp4_kppv(DataA,DataT,labelA,labelT,K,ListeClass)

Entrées :

DataA : (Array{Float64,2}) données d'apprentissage

DataT : (Array{Float64,2}) données de test

labelA : (Array{Float64,2}) labels d'apprentissage

labelT : (Array{Float64,2}) labels de test

K : (Int64) nombre de voisins à prendre en compte

ListeClass : (Array{Intt64,2}) l'ensemble des classes

Nt_test : (Int64) le nombre de test

Sorties:

Partition : (Array{Intt64,2}) les classes des Nt_test premiers éléments de DataT

confusion : (Array{Float64,2}) la matrice de confusion

nb_erreurs : (Int64) nombre d'erreurs effectuées lors de la classification